// src/generated/models/PowerPlatformforAdminsV2Model.ts
export type CrossTenantConnectionconnectionType = 'Inbound'|'Outbound';

export interface CrossTenantConnection {
  // The Azure AD tenant ID to or from which the cross-tenant connection occurred.
  tenantId: string;
  // The direction of the cross-tenant connection.
  connectionType: CrossTenantConnectionconnectionType;
}

export type CrossTenantConnectionReportstatus = 'Received'|'InProgress'|'Completed'|'Failed';

export interface CrossTenantConnectionReport {
  // The Azure AD tenant ID for which the report was generated.
  tenantId: string;
  // The report ID.
  reportId: string;
  // The date when the cross-tenant connection report was requested.
  requestDate: string;
  // The start of the report date window.
  startDate: string;
  // The end of the report date window.
  endDate: string;
  status: CrossTenantConnectionReportstatus;
  // The page of cross-tenant connections occurring within the report date window.
  connections?: CrossTenantConnection[];
  // Next page URI for additional cross-tenant connections.
  "@odata.nextLink"?: string;
}
// ...
export interface EnvironmentGroup {
  id?: string;
  displayName?: string;
  description?: string;
  parentGroupId?: string;
  childrenGroupIds?: string[];
  createdTime?: string;
  createdBy?: Principal;
  lastModifiedTime?: string;
  lastModifiedBy?: Principal;
}
// ...
export interface Principal {
  id?: string;
  displayName?: string;
  email?: string;
  type?: string;
  tenantId?: string;
  userPrincipalName?: string;
}